refactor: dedicated package for beans with `@Produces` methods
- Move ApplicationInfo.java to a new package `about`
- Rename `PaperTrailSdkBeanProvider.java` to `PaperTrailSdkProducer.java` and move to `producers` package
